# apps/admin-api/src/support/api-docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/mails/bulk/raw": {
      "post": {
        "description": "관리자가 자유 입력한 subject/html 으로 다수 수신자에게 발송합니다.\n개별 수신자별 성공/실패가 응답에 포함됩니다 (전체가 실패해도 200).\n최대 100명까지 한 번에 발송할 수 있습니다.\n\n**Required Permissions (ALL):** `MAIL_SEND` (고객에게 이메일 발송)",
        "operationId": "AdminMailController_sendBulkRaw",
        "parameters": [
          {
            "required": true,
            "name": "x-user-lang",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "en-US",
                "ko",
                "fr",
                "es",
                "de",
                "ja",
                "ms"
              ]
            },
            "description": "사용자의 언어 코드"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSendBulkRawReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "SUCCESS"
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminSendMailRes"
                    },
                    "error": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "유효하지 않은 수신자",
            "content": {
              "application/json": {
                "examples": {
                  "example1": {
                    "summary": "유효하지 않은 수신자",
                    "value": {
                      "result": "ERROR",
                      "data": null,
                      "error": {
                        "code": "INVALID_MAIL_RECIPIENT",
                        "message": "유효하지 않은 수신자",
                        "data": null
                      }
                    }
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "ERROR"
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "$ref": "#/components/schemas/AdminApiErrorMessage"
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Access JWT": []
          }
        ],
        "summary": "메일 발송 (raw)",
        "tags": [
          "Mail"
        ]
      }
    },
    "/v1/mails/bulk/template": {
      "post": {
        "description": "사전 정의된 템플릿 + 변수로 다수 수신자에게 발송합니다.\ntemplateId 와 그에 맞는 vars 를 입력합니다 (templateId 별 필요 키는 vars 필드 설명 참고).\n\n**Required Permissions (ALL):** `MAIL_SEND` (고객에게 이메일 발송)",
        "operationId": "AdminMailController_sendBulkByTemplate",
        "parameters": [
          {
            "required": true,
            "name": "x-user-lang",
            "in": "header",
            "schema": {
              "type": "string",
              "enum": [
                "en-US",
                "ko",
                "fr",
                "es",
                "de",
                "ja",
                "ms"
              ]
            },
            "description": "사용자의 언어 코드"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSendBulkTemplateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "SUCCESS"
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/AdminSendMailRes"
                    },
                    "error": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "유효하지 않은 수신자",
            "content": {
              "application/json": {
                "examples": {
                  "example1": {
                    "summary": "유효하지 않은 수신자",
                    "value": {
                      "result": "ERROR",
                      "data": null,
                      "error": {
                        "code": "INVALID_MAIL_RECIPIENT",
                        "message": "유효하지 않은 수신자",
                        "data": null
                      }
                    }
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "enum": [
                        "ERROR"
                      ]
                    },
                    "data": {
                      "type": "null",
                      "nullable": true
                    },
                    "error": {
                      "$ref": "#/components/schemas/AdminApiErrorMessage"
                    }
                  },
                  "required": [
                    "result",
                    "data",
                    "error"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "Access JWT": []
          }
        ],
        "summary": "메일 발송 (template)",
        "tags": [
          "Mail"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nx Nest Starter Admin API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Access JWT": {
        "scheme": "bearer",
        "name": "Access JWT",
        "type": "http",
        "in": "header",
        "description": "Access JWT for user authentication",
        "x-apidog": {
          "authType": "bearer",
          "addTokenTo": "header",
          "headerPrefix": "Bearer",
          "token": "{{bearerToken}}"
        }
      }
    },
    "schemas": {
      "AdminApiErrorMessage": {
        "type": "object",
        "properties": {
          "traceId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "traceId",
          "code",
          "message"
        ]
      },
...
      "FailedRecipient": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "reason": {
            "type": "string",
            "description": "SES 호출 실패 사유"
          }
        },
        "required": [
          "email",
          "reason"
        ]
      },
      "AdminSendMailRes": {
        "type": "object",
        "properties": {
          "successCount": {
            "type": "integer",
            "description": "발송 성공 건수"
          },
          "failedCount": {
            "type": "integer",
            "description": "발송 실패 건수"
          },
          "success": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "description": "발송 성공한 수신자 이메일"
          },
          "failed": {
            "description": "발송 실패한 수신자와 사유",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FailedRecipient"
            }
          }
        },
        "required": [
          "successCount",
          "failedCount",
          "success",
          "failed"
        ]
      },
      "AdminSendBulkRawReq": {
        "type": "object",
        "properties": {
          "recipients": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "minItems": 1,
            "maxItems": 100,
            "description": "수신자 이메일 목록 (최대 100건)",
            "example": [
              "user1@example.com",
              "user2@example.com"
            ]
          },
          "subject": {
            "type": "string",
            "maxLength": 200,
            "description": "메일 제목"
          },
          "html": {
            "type": "string",
            "description": "메일 본문 HTML. admin 이 직접 작성하므로 sanitize 책임은 admin 에게 있음."
          },
          "lang": {
            "type": "string",
            "enum": [
              "en-US",
              "ko",
              "fr",
              "es",
              "de",
              "ja",
              "ms"
            ],
            "default": "en-US",
            "description": "발송 언어 코드. 미지정 시 en-US. (이 호출의 모든 수신자에게 동일 언어로 발송)"
          }
        },
        "required": [
          "recipients",
          "subject",
          "html"
        ]
      },
      "AdminSendBulkTemplateReq": {
        "type": "object",
        "properties": {
          "recipients": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "email"
            },
            "minItems": 1,
            "maxItems": 100,
            "description": "수신자 이메일 목록 (최대 100건)"
          },
          "templateId": {
            "type": "string",
            "enum": [
              "announcement"
            ],
            "description": "템플릿 식별자",
            "example": "announcement"
          },
          "vars": {
            "type": "object",
            "additionalProperties": true,
            "description": "템플릿이 요구하는 변수 객체. templateId 별 필요 키:\n- `announcement`: { subject: string, bodyHtml: string }",
            "example": {
              "subject": "점검 안내",
              "bodyHtml": "<p>3월 5일 02:00~04:00 시스템 점검 예정입니다.</p>"
            }
          },
          "lang": {
            "type": "string",
            "enum": [
              "en-US",
              "ko",
              "fr",
              "es",
              "de",
              "ja",
              "ms"
            ],
            "default": "en-US",
            "description": "발송 언어 코드. 미지정 시 en-US. (이 호출의 모든 수신자에게 동일 언어로 발송)"
          }
        },
        "required": [
          "recipients",
          "templateId",
          "vars"
        ]
      },
...
    }
  }
}